refactor(bot): MatchTrackingServiceへ監視処理を集約
watcher処理とtick内キャッシュがmatch_tracking.tsに集中しており、API clientやDiscord clientなどのmodule-level依存を分離しづらい状態だった。

createMatchTrackingServiceを導入し、Backend resource、notifier、renderer、clock、logger、configを注入する構成へ移した。match_tracking.tsは環境変数読み取りとDiscord向けassembly、worker管理だけを担うFacadeにする。

1 tick内のaccount / active game / match取得共有、通知統合、一部watcher失敗時の後続処理継続は既存回帰テストで維持し、service単体テストで例外後の継続処理を固定した。
